Remove timeout handling from cmdstalk
This removes the timeout handling behaviour from cmdstalk. Currently
if a task times out in cmdstalk, beanstalk will immediately make it
ready to run and when cmdstalk sees that it has timed out once before
it will immediately bury it, preventing it from ever being run again.

This drops that behaviour, so that if a timed out task is seen, it
lets it continue to execute as normal. This does mean there's a risk
of the same task being executed more than once simultaneously, but
that has been the reality for contests for a very long time and seems
to be have managed fine.

Other possible solutions are:
 - When seeing a task that's timed out, immediately release it with an
 exponential timeout until releases >= timeouts. Could result in some
 weird behaviour though; beanstalk stats aren't fantastic for controlling
 this kind of behaviour.
 - ???